[fswatch\src\windows\getopt.c] Use `_dupenv_s` instead of `getenv` on MSVC; [fswatch\src\fswatch.cpp] Print error; [fswatch\src\libfswatch\c\cevent.h] Use `size_t` instead of `unsigned int` to match expected type; [fswatch\src\libfswatch\c++\windows\win_directory_change_event.cpp] Cast types; deduplicate mask; [fswatch\src\libfswatch\filter.cpp] Use `size_t` for index and length; [fswatch\src\libfswatch\monitor.cpp] print error; [fswatch\test\src\fswatch.c] Use `size_t` for `event_num`

+++ libfswatch/src/libfswatch/c/cevent.h
@@ -135,7 +135,7 @@ extern "C"
    * a pointer to it.
    */
   typedef void (*FSW_CEVENT_CALLBACK)(fsw_cevent const *const events,
-                                      const unsigned int event_num,
+                                      const size_t event_num,
                                       void *data);
 
 #  ifdef __cplusplus
